Add a moveable 'player' in playing state
This is the first in a series of commits that will hopefully exemplify
how gameplay could work. I put some actual data in the Game class (the
player's direction, x position and y position) and that data gets
updated via movement commands that are bound to the arrow buttons. There
are no constraints yet, so you can move it completely off the screen if
you want, but I'll cover that soon.