Added the first unit test
This was much more of a pain than I thought it would be. I finally got it working after removing the pre-compiled headers, removing a couple linker files that didn't actually exist, adding a pre-linker step to compile all the obj files into a lib file, and referencing that lib file.