Allow command args to be passed with game commands
I could've broken this out into more commits, but got a little carried away. This allows us to pass args along with a game command, in this example we pass the direction the player wants to move as an argument to a generic 'MovePlayer' command. It may seem like overkill for what we're using it for now, but will definitely be necessary once commands get more complicated.