Fixed a little glitch with the clock
This was a tough one, I noticed all of a sudden that the diagnostics were reporting that roughly half of all frames were lag frames, yet the game was still going very fast. I'm still not 100% sure why, but setting the next frame's start time using the actual clock seemed to fix it.