Refactor how the clock handles frames
It occurred to me that the clock isn't really set up properly, because it shouldn't have to know exactly when to set the frame start time (and it shouldn't be done at the end of the last frame anyway). We should be able to tell it when to start the timer for the frame, and then tell it when to wait for the frame to end. This doesn't really affect gameplay at the moment, but I could see it becoming a potential problem later, so it's better to address it now.