Separate arena size types
This is gonna be a little messy at first, I realized there will need to be some separation between units of size when drawing the arena vs. moving the player around the arena. When rendering, one 'unit' is one character, but in the game logic we should have the notion of 'game units'. Right now they're one-to-one, just to keep the peace, but I'll need to add some kind of translation later on.