Now targets Windows subsystem, creates its own console window on launch
This means that we ALWAYS get a conhost window that we can control
(specifically, resize!). If you F5 in VS, it'll launch the "Windows
applicaiton," which creates no UI by default, so it's natural when it
creates its own console window. Likewise, it behaves identically when
running from Windows Terminal--a new conhost window is created, rather
than running it directly in that Terminal tab.